+++ tests/ci.bash
@@ -27,7 +27,4 @@ if [ ! -z $CI_RUN ]; then
         echo "";
         
     fi
-    sudo apt-get -y install python-all-dev python-setuptools;
-    sudo easy_install -U setuptools pip;
-    sudo pip install --upgrade pip pytz;
 fi